refactor: organize transaction modal files by responsibility
