+++ frontend/src/transactions/components/TransactionModal/Shell.tsx
@@ -2,7 +2,7 @@ import type { FormEvent, ReactNode } from 'react'
 import { createPortal } from 'react-dom'
 import { AnimatePresence, motion } from 'motion/react'
 import { ReceiptText, X } from 'lucide-react'
-import { EASE } from '@/transactions/components/TransactionModal/transactionModalConstants'
+import { EASE } from '@/transactions/components/TransactionModal/constants'
 
 interface TransactionModalShellProps {
   open: boolean
