+++ frontend/src/transactions/components/TransactionModal/utils/initialForm.ts
@@ -2,12 +2,12 @@ import type { AccountsOverview } from '@/api/accounts'
 import type { Category } from '@/api/categories'
 import type { Currency } from '@/api/currency'
 import type { Transaction } from '@/api/transactions'
-import { INITIAL_TRANSACTION_FORM } from '@/transactions/components/TransactionModal/transactionModalConstants'
-import { amountToInputString, getTodayLocalDateInputValue } from '@/transactions/components/TransactionModal/transactionModalMoney'
+import { INITIAL_TRANSACTION_FORM } from '@/transactions/components/TransactionModal/constants'
+import { amountToInputString, getTodayLocalDateInputValue } from '@/transactions/components/TransactionModal/utils/money'
 import type {
   TransactionFormValues,
   TransactionModalKind,
-} from '@/transactions/components/TransactionModal/transactionModalTypes'
+} from '@/transactions/components/TransactionModal/types'
 
 interface BuildInitialTransactionFormOptions {
   transaction?: Transaction
